+++ phenodata/command.py
@@ -8,7 +8,7 @@ from phenodata import __version__
 from phenodata.ftp import FTPSession
 from phenodata.dwd.cdc import DwdCdcClient
 from phenodata.dwd.pheno import DwdPhenoData, DwdPhenoDataHumanizer
-from phenodata.util import boot_logging, normalize_options
+from phenodata.util import boot_logging, normalize_options, options_convert_lists
 
 """
 phenodata is a data acquisition and manipulation toolkit for open access phenology data.
@@ -31,7 +31,7 @@ def run():
       phenodata list-quality-bytes --source=dwd [--format=csv]
       phenodata list-filenames --source=dwd --dataset=immediate --partition=recent [--filename=Hasel,Schneegloeckchen] [--year=2017]
       phenodata list-urls --source=dwd --dataset=immediate --partition=recent [--filename=Hasel,Schneegloeckchen] [--year=2017]
-      phenodata (observations|forecast) --source=dwd --dataset=immediate --partition=recent [--filename=Hasel,Schneegloeckchen] [--station-id=164,717] [--species-id=113,127] [--phase-id=5] [--quality-level=10] [--quality-byte=1,2,3] [--station=berlin,brandenburg] [--species=hazel,snowdrop] [--phase=flowering] [--quality=blubb] [--year=2017] [--humanize] [--show-ids] [--language=german] [--long-station] [--sort=Datum] [--format=csv]
+      phenodata (observations|forecast) --source=dwd --dataset=immediate --partition=recent [--filename=Hasel,Schneegloeckchen] [--station-id=164,717] [--species-id=113,127] [--phase-id=5] [--quality-level=10] [--quality-byte=1,2,3] [--station=berlin,brandenburg] [--species=hazel,snowdrop] [--species-preset=mellifera-primary] [--phase=flowering] [--quality=ROUTKLI] [--year=2017] [--humanize] [--show-ids] [--language=german] [--long-station] [--sort=Datum] [--format=csv]
       phenodata --version
       phenodata (-h | --help)
 
@@ -51,6 +51,7 @@ def run():
       --station=<station>       Filter by strings from "stations" data (comma-separated list)
       --species=<species>       Filter by strings from "species" data (comma-separated list)
       --phase=<phase>           Filter by strings from "phases" data (comma-separated list)
+      --species-preset=<preset> Filter by strings from "species" data (comma-separated list) loaded from ``presets.json`` file
 
     Data output options:
       --format=<format>         Output data in designated format. Choose one of "tabular", "json", "csv" or "string".
@@ -76,7 +77,15 @@ def run():
     boot_logging(options)
 
     # Normalize commandline options
-    options = normalize_options(options, encoding='utf-8', list_items=[
+    options = normalize_options(options, encoding='utf-8')
+
+    # Expand options
+    preset_name = options['species-preset']
+    if preset_name:
+        options['species'] = DwdPhenoData.load_preset('options', 'species', preset_name)
+
+    # Coerce comma-separated list fields
+    options_convert_lists(options, list_items=[
 
         # Acquisition parameters
         'filename',
@@ -117,7 +126,9 @@ def run():
         humanizer = DwdPhenoDataHumanizer(language=options['language'], long_station=options['long-station'], show_ids=options['show-ids'])
         client = DwdPhenoData(cdc=cdc_client, humanizer=humanizer, dataset=options.get('dataset'))
     else:
-        raise DocoptExit('Data source "{}" not implemented'.format(options['source']))
+        message = 'Data source "{}" not implemented'.format(options['source'])
+        logger.error(message)
+        raise DocoptExit(message)
 
     # Dispatch command
     data = None