+++ CHANGES.rst
@@ -8,6 +8,9 @@ development
 - More compact date output for "tabular" mode
 - Fix datetime coercion when encountering invalid datetime values
 - Humanize searching in observation and forecast data for stations, species, phases and quality information
+- Implement predefined sets of parameters using ``presets.json``.
+  Apply with e.g. "--species-preset=mellifera-de-primary".
+  See also https://community.hiveeyes.org/t/phanologischer-kalender-entwicklung/664/23.
 
 2018-03-14 0.5.0
 ================